Support install-only workflow.
This commit adds support for an install-only workflow.
If command is not specified, this will halt execution
after the CLI is installed.

This attempts to replicate the @pulumi/setup-pulumi action.

To test this, most logic in main() was moved to a separate file which
can be imported in a test context, unlike main.ts, which executes
main() during resolution.
