+++ src/__tests__/run.test.ts
@@ -3,6 +3,53 @@ import { login } from '../login';
 
 const spy = jest.spyOn(pulumiCli, 'run');
 
+const installConfig: Record<string, string> = {
+  command: undefined,
+  pulumiVersion: "^4",
+};
+
+describe('Config without a provided command', () => {
+  let oldWorkspace = '';
+  beforeEach(() => {
+    spy.mockClear();
+    jest.resetModules();
+    // Save, then restore the current env var for GITHUB_WORKSPACE 
+    oldWorkspace = process.env.GITHUB_WORKSPACE;
+    process.env.GITHUB_WORKSPACE = 'n/a';
+  });
+  afterEach(() => {
+    process.env.GITHUB_WORKSPACE = oldWorkspace;
+  });
+
+  it('should not be validated by makeConfig', async () => {
+    jest.mock('@actions/core', () => ({
+      getInput: jest.fn((name: string) => {
+        return installConfig[name];
+      }),
+      info: jest.fn(),
+    }));
+    jest.mock('@actions/github', () => ({
+      context: {},
+    }));
+    const { makeConfig } = require('../config');
+    await expect(makeConfig())
+      .rejects
+      .toThrow();
+  });
+
+  it('should be validated by makeInstallationConfig', async() => {
+    jest.mock('@actions/core', () => ({
+      getInput: jest.fn((name: string) => {
+        return installConfig[name];
+      }),
+      info: jest.fn(),
+    }));
+    const { makeInstallationConfig } = require('../config');
+    const conf = await makeInstallationConfig()
+    expect(conf).toBeTruthy();
+  });
+});
+
 describe('main.login', () => {
   beforeEach(() => {
     spy.mockClear();
