+++ src/main.ts
@@ -1,112 +1,25 @@
-import { resolve } from 'path';
 import * as core from '@actions/core';
-import {
-  ConfigMap,
-  LocalProgramArgs,
-  LocalWorkspace,
-  LocalWorkspaceOptions,
-} from '@pulumi/pulumi/automation';
-import invariant from 'ts-invariant';
-import YAML from 'yaml';
-import { Commands, makeConfig } from './config';
-import { environmentVariables } from './libs/envs';
-import { handlePullRequestMessage } from './libs/pr';
-import * as pulumiCli from './libs/pulumi-cli';
-import { login } from './login';
+import { makeConfig, makeInstallationConfig } from './config';
+import { installOnly, runAction } from './run';
 
 const main = async () => {
-  const config = await makeConfig();
-  core.debug('Configuration is loaded');
-
-  await pulumiCli.downloadCli(config.options.pulumiVersion);
-  await login(config.cloudUrl, environmentVariables.PULUMI_ACCESS_TOKEN);
-
-  const workDir = resolve(
-    environmentVariables.GITHUB_WORKSPACE,
-    config.workDir,
-  );
-  core.debug(`Working directory resolved at ${workDir}`);
-
-  const stackArgs: LocalProgramArgs = {
-    stackName: config.stackName,
-    workDir: workDir,
-  };
-
-  const stackOpts: LocalWorkspaceOptions = {};
-  if (config.secretsProvider != '') {
-    stackOpts.secretsProvider = config.secretsProvider;
-  }
-
-  const stack = await (config.upsert
-    ? LocalWorkspace.createOrSelectStack(stackArgs, stackOpts)
-    : LocalWorkspace.selectStack(stackArgs, stackOpts));
-
-  const projectSettings = await stack.workspace.projectSettings();
-  const projectName = projectSettings.name;
-
-  const onOutput = (msg: string) => {
-    core.debug(msg);
-    core.info(msg);
-  };
-
-  if (config.configMap != '') {
-    const configMap: ConfigMap = YAML.parse(config.configMap);
-    await stack.setAllConfig(configMap);
-  }
-
-  if (config.refresh) {
-    core.startGroup(`Refresh stack on ${config.stackName}`);
-    await stack.refresh({ onOutput });
-    core.endGroup();
-  }
-
-  core.startGroup(`pulumi ${config.command} on ${config.stackName}`);
-
-  const actions: Record<Commands, () => Promise<string>> = {
-    up: () => stack.up({ onOutput, ...config.options }).then((r) => r.stdout),
-    update: () =>
-      stack.up({ onOutput, ...config.options }).then((r) => r.stdout),
-    refresh: () =>
-      stack.refresh({ onOutput, ...config.options }).then((r) => r.stdout),
-    destroy: () =>
-      stack.destroy({ onOutput, ...config.options }).then((r) => r.stdout),
-    preview: async () => {
-      const { stdout, stderr } = await stack.preview(config.options);
-      onOutput(stdout);
-      onOutput(stderr);
-      return stdout;
-    },
-  };
-
-  core.debug(`Running action ${config.command}`);
-  const output = await actions[config.command]();
-  core.debug(`Done running action ${config.command}`);
-
-  core.setOutput('output', output);
-
-  const outputs = await stack.outputs();
-
-  for (const [outKey, outExport] of Object.entries(outputs)) {
-    core.setOutput(outKey, outExport.value);
-    if (outExport.secret) {
-      core.setSecret(outExport.value);
+  try {
+    const downloadConfig = await makeInstallationConfig();
+    installOnly(downloadConfig);
+    core.info("Pulumi has been successfully installed. Exiting.")
+  } catch(error) {
+    // This could be one of two error types:
+    // 1. Validation failed, which is a recoverable error.
+    // 2. Installation failed, which is unrecoverable.
+    // We check the error type to see whether to procede or bail.
+    if(error?.name === "ValidationError") {
+      const config = await makeConfig();
+      core.debug('Configuration is loaded');
+      runAction(config);
+    } else {
+      throw(error);
     }
   }
-
-  if (
-    config.commentOnPrNumber ||
-    (config.commentOnPr && config.isPullRequest)
-  ) {
-    core.debug(`Commenting on pull request`);
-    invariant(config.githubToken, 'github-token is missing.');
-    handlePullRequestMessage(config, projectName, output);
-  }
-
-  if (config.remove && config.command === 'destroy') {
-    stack.workspace.removeStack(stack.name);
-  }
-
-  core.endGroup();
 };
 
 (async () => {
